Refactor: relocate useFocusTrap to _internal/overlay/
Prep step for the Drawer component which will share the same hook.
No behavior change. Modal's Content.tsx imports from the new path.
