+++ packages/design-system/src/components/Modal/Content.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useLayoutEffect, type CSSProperties, type ReactNode } from 'react';
 import clsx from 'clsx';
 import { useModalContext } from './context';
-import { useFocusTrap } from './useFocusTrap';
+import { useFocusTrap } from '../_internal/overlay/useFocusTrap';
 import { modalStack } from './useModalStack';
 import styles from './Modal.module.scss';
